requirements.txt
africastalking==1.2.4
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.1
contextlib2==21.6.0
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
idna==3.3
pytz==2022.2.1
requests==2.28.1
schema==0.7.5
sqlparse==0.4.2
urllib3==1.26.12
